Image Subtraction ("Imsub")

+++ winterdrp/processors/astromatic/sextractor/sextractor.py
@@ -1,10 +1,12 @@
 import os
+import shutil
+
 import numpy as np
 import logging
 import astropy.io.fits
 from winterdrp.processors.astromatic.sextractor.sourceextractor import run_sextractor_single, default_saturation, \
     run_sextractor_dual
-from winterdrp.processors.base_processor import BaseProcessor
+from winterdrp.processors.base_processor import BaseImageProcessor
 from winterdrp.paths import get_output_dir, get_temp_path, latest_mask_save_key
 
 logger = logging.getLogger(__name__)
@@ -12,7 +14,7 @@ logger = logging.getLogger(__name__)
 sextractor_header_key = 'SRCCAT'
 
 
-class Sextractor(BaseProcessor):
+class Sextractor(BaseImageProcessor):
     base_key = "sextractor"
 
     def __init__(
@@ -28,6 +30,8 @@ class Sextractor(BaseProcessor):
             checkimage_type: str | list = None,
             gain: float = None,
             dual: bool = False,
+            cache: bool = False,
+            mag_zp :float = None,
             *args,
             **kwargs
     ):
@@ -44,6 +48,8 @@ class Sextractor(BaseProcessor):
         self.checkimage_type = checkimage_type
         self.gain = gain
         self.dual = dual
+        self.cache = cache
+        self.mag_zp = mag_zp
 
     def get_sextractor_output_dir(self):
         return get_output_dir(self.output_sub_dir, self.night_sub_dir)
@@ -67,6 +73,9 @@ class Sextractor(BaseProcessor):
             data = images[i]
 
             det_image, measure_image, det_header, measure_header = None, None, None, None
+            if self.gain is None:
+                if 'GAIN' in header.keys():
+                    self.gain = header['GAIN']
             if self.dual:
                 det_header = headers[i][0]
                 measure_header = headers[i][1]
@@ -77,16 +86,22 @@ class Sextractor(BaseProcessor):
                 self.gain = measure_header["GAIN"]
 
             temp_path = get_temp_path(sextractor_out_dir, header["BASENAME"])
+            if not os.path.exists(temp_path):
+                self.save_fits(data, header, temp_path)
             temp_files = [temp_path]
 
             mask_path = None
             if latest_mask_save_key in header.keys():
-                mask_path = get_temp_path(sextractor_out_dir, header[latest_mask_save_key])
-                if not os.path.exists(mask_path):
+                image_mask_path = os.path.join(sextractor_out_dir, header[latest_mask_save_key])
+                temp_mask_path = get_temp_path(sextractor_out_dir, header[latest_mask_save_key])
+                if os.path.exists(image_mask_path):
+                    shutil.copyfile(image_mask_path,temp_mask_path)
+                    mask_path = temp_mask_path
+                    temp_files.append(mask_path)
+                else:
                     mask_path = None
 
             if mask_path is None:
-                self.save_fits(data, header, temp_path)
                 mask_path = self.save_mask(data, header, temp_path)
                 temp_files.append(mask_path)
             output_cat = os.path.join(sextractor_out_dir, header["BASENAME"].replace(".fits", ".cat"))
@@ -123,12 +138,15 @@ class Sextractor(BaseProcessor):
                     checkimage_name=self.checkimage_name,
                     checkimage_type=self.checkimage_type,
                     gain=self.gain,
-                    catalog_name=output_cat
+                    catalog_name=output_cat,
+                    mag_zp=self.mag_zp
                 )
 
-            for temp_file in temp_files:
-                os.remove(temp_file)
-                logger.info(f"Deleted temporary file {temp_file}")
+            logger.info(f'Cache save is {self.cache}')
+            if not self.cache:
+                for temp_file in temp_files:
+                    os.remove(temp_file)
+                    logger.info(f"Deleted temporary file {temp_file}")
 
             header[sextractor_header_key] = os.path.join(sextractor_out_dir, output_cat)
